Implement mapping for champseason filters
