+++ src/iRLeagueApiCore.Server/Handlers/HandlerBase.cs
@@ -1,4 +1,5 @@
-﻿using iRLeagueApiCore.Server.Models;
+﻿using iRLeagueApiCore.Common.Models;
+using iRLeagueApiCore.Server.Models;
 using System.Diagnostics.CodeAnalysis;
 
 namespace iRLeagueApiCore.Server.Handlers;
@@ -160,5 +161,47 @@ public abstract class HandlerBase<THandler, TRequest>
             mappedResults.Add(sessionResult);
         }
         return mappedResults;
+    }
+
+    protected async Task<ICollection<FilterOptionEntity>> MapToFilterOptionListAsync(LeagueUser user, IEnumerable<ResultFilterModel> filterModels,
+        ICollection<FilterOptionEntity> filterEntities, CancellationToken cancellationToken)
+    {
+        foreach (var filterModel in filterModels)
+        {
+            var filterOptionEntity = filterModel.FilterOptionId == 0 ? null : filterEntities
+                .FirstOrDefault(x => x.FilterOptionId == filterModel.FilterOptionId);
+            if (filterOptionEntity is null)
+            {
+                filterOptionEntity = CreateVersionEntity(user, new FilterOptionEntity());
+                filterEntities.Add(filterOptionEntity);
+                filterOptionEntity.LeagueId = dbContext.LeagueProvider.LeagueId;
+            }
+            await MapToFilterOptionEntityAsync(user, filterModel, filterOptionEntity, cancellationToken);
+        }
+        var deleteFilterEntities = filterEntities
+            .Where(x => filterModels.Any(y => y.FilterOptionId == x.FilterOptionId) == false);
+        foreach (var deleteFilterEntity in deleteFilterEntities)
+        {
+            filterEntities.Remove(deleteFilterEntity);
+        }
+        return filterEntities;
+    }
+
+    protected Task<FilterOptionEntity> MapToFilterOptionEntityAsync(LeagueUser user, ResultFilterModel filterModel, FilterOptionEntity filterOptionEntity,
+        CancellationToken cancellationToken)
+    {
+        var condition = filterOptionEntity.Conditions.FirstOrDefault();
+        if (condition is null)
+        {
+            condition = new FilterConditionModel();
+            filterOptionEntity.Conditions.Add(condition);
+        }
+        condition.Comparator = filterModel.Condition.Comparator;
+        condition.FilterType = filterModel.Condition.FilterType;
+        condition.ColumnPropertyName = filterModel.Condition.ColumnPropertyName;
+        condition.FilterValues = filterModel.Condition.FilterValues;
+        condition.Action = filterModel.Condition.Action;
+        UpdateVersionEntity(user, filterOptionEntity);
+        return Task.FromResult(filterOptionEntity);
     }
 }
