+++ src/iRLeagueApiCore.Services/iRLeagueApiCore.Services.csproj
@@ -17,8 +17,8 @@
 	</ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="iRLeagueApiCore.Common" Version="0.9.0-dev.1" />
-    <PackageReference Include="iRLeagueDatabaseCore" Version="0.9.0-dev.1" />
+    <PackageReference Include="iRLeagueApiCore.Common" Version="0.9.0-dev.2" />
+    <PackageReference Include="iRLeagueDatabaseCore" Version="0.9.0-dev.2" />
     <PackageReference Include="Microsoft.AspNetCore.Http.Abstractions" Version="2.2.0" />
     <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="6.0.0" />
     <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="6.0.0" />
