+++ test/iRLeagueApiCore.UnitTests/iRLeagueApiCore.UnitTests.csproj
@@ -21,8 +21,8 @@
     <PackageReference Include="AutoFixture.AutoMoq" Version="4.17.0" />
     <PackageReference Include="FluentAssertions" Version="6.7.0" />
     <PackageReference Include="FluentIdentityBuilder" Version="0.0.3" />
-    <PackageReference Include="iRLeagueApiCore.Common" Version="0.9.0-dev.1" />
-    <PackageReference Include="iRLeagueDatabaseCore" Version="0.9.0-dev.1" />
+    <PackageReference Include="iRLeagueApiCore.Common" Version="0.9.0-dev.2" />
+    <PackageReference Include="iRLeagueDatabaseCore" Version="0.9.0-dev.2" />
     <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="6.0.9" />
     <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="6.0.9">
       <PrivateAssets>all</PrivateAssets>
